Added functions rescale, data_distributions, removed lower limit of results.
The rescale function was added to neural_network.py. This function scales each feature present in the data by subtracting the mean of the feature in the dataset and dividing by the standard deviation. This scaling occurs only if the keyword argument scale = True. If the keyword argument scale=False then the inverse transformation is performed; meaning that each feature will be first multiplied by their standard deviation across the dataset, to whoms value the mean will then be added. The implicit assumption made in this implementation, is that the datasets passed into the rescale function are of sufficient size that the standard deviation and mean do not vary a significant amount for different datasets passed to the function.

The rescale function was called once in at the end of _parse_features found in neural_network.py and once in _simulate, in th-e-simulation.py, following the prediction of pv_power and the concatenation of the relevant target value to this result (line: 156). It is imperative that the rescale of the result, occurs before the calculation of the error. This is because the error does not transform like the features do, so applying the inverse scaling trafo to the error will not result in the correct error values.

Ex: [(x_{1}-m)/s -(x_{2}-m)/s]*s+m = x_{1}-x_{2}+m != x_{1}-x_{2}

The data_distribution function creates a distribution folder with the sub_folders 'raw' and 'scale' in the model sub-folder. Histograms displaying the distributions of the various features are saved in the folder 'raw' if the keyword argument 'scale', accepted by the data_distribution function, is equal to the bool False. On the otherhand, if the keyword argument 'scale' is equal to the bool True then the histograms are saved in the subfolder 'scale'. For this reason in _parse_features we call the function data_distributions with the keyword argument scale=False directly before the rescale function is called, and then we call the same function once more with the keyword argument scale=True directly after the rescale function. In this way the distributions of the various features is visualized, and the effect that the scaling function has on these distributions can be seen directly seen by comparing the results in the 'raw' and 'scale' folders.

In the function _run_step similarly found in neural_network.py the lower limit placed to supress values smaller then 1e-3 was removed as feature scaling potentially makes the region from 0 to 1e-3 relevant to our results. The possibility still remains to supress irrelevently small values in the future, but this value must be chosen with respect to the scale of our pv_power values. In the case of working with data transformed with the rescale function, this value is likely lower then 1e-3.