split rescale function into gen_trafo and rescale function, move distribution dir out of model dir
The changes made have an overall desire for simpler and more readable code and a more robust technique of telling the function which features should be scaled by the same function.

Previously, the rescale function generated the scaling function and then applied that function to our data. These are clearly two seperate functional tasks, and it was thus decided to ensure that the rescale function, itself, was only composed of a few lines of code, deciding to which features preexisting scaling functions should be applied. The task of generating the functions, that rescale ultimately uses, was pushed over to the gen_trafo function, whose output is ultimately stored in the dictionary self.trafos. as well as having the following in mind:

In order for a function to transform the features we want to be scaled the in the same way, the same, the function has to be told in some way that these features belong to the same 'scaling group'. One solution is to force the user to manually group the features which have to be scaled the same in the forecast.cfg file. Another solution is to agree upon a naming convention in which all features, which should be scaled by the same function, is somehow indicated in the features name (start with pv if scaled like pv_power).

I chose the first option because it is the most transparent (easiest for an outside user to understand). The only difficulty lies in remembering to insert variables in the scaling list, which only arise in the output, along with variables which are to be scaled in the input (e.g. 'pv_power_est')