Reference model in log messages being simulated, added custom metric to tensorboard.
The only changes made in this commit was adding a reference to the model being simulated in the logging messages, so it is easier to track the simulation as it runs. Furthermore a custom metric function which tracks the norm loss was added immediately following the call of model.fit. This should allow the comparison of both scaled and unscaled models.