Added new network classes and other changes to allow the simulation of consecutive models.
First and foremost the forecast.py script had to be changes so as to initiate the correct class depending on the forecast.cfg file. This allows the network class being used (e.g. ConvLSTM, MultiLayerPerceptron) to be controlled within the config file, which ultimately allows different network architectures to be simulated simultaneously.

The other big change was to add a NeuralNetwork subclass Multilayerperceptron. This is just a standard MLP network and varies from ConvLSTM only in the contents of its build function (necessary layers and units required for this class were imported at the top of neural_network.py).

Small changes carried out in this commit are detailed here. The name dependency of the retrievel of configs with cls.__name__ was removed: this dependency required the section in the config file to be relabled to correspond to the class name everytime a new neural network subclass was used in order to forecast the pv power output. In place of this dependency the corresponding section in forecast.cfg was renamed to a more generic label 'NeuralNetwork'. Whatever configurations are in this section are loaded into the configs parser in neural_network.py. The queue folder was added to .gitignore.

Added queue to .gitignore, loss from mae to logcosh, allow configuration of network structure in configs, remove network name depen