Fixed bug in forecast, edited NN classes so parameters are mostly set in configurations files.
if model is ... was not working upon testing. is was thus changed to "==". Finally most parameters appearing  in the build functions of the various NN classes were add to the forecast configs where possible, and then retrieved from configs in the class. This allows the complete configuration of the network from this file, and thus the ability to efficiently test different network structures.