Requisite modifications to allow the output of the weight parameters in the model_comparison.xlsx:
It was decided to write the number of network weights to the xlsx output doc, so as to allow a comparison of 'network complexity' not simply based on the runtime of the simulations; such a comparison is given to be flawed, especially if other processes are being carried out by the cpu as th_e_simulation.py runs. This should allow more objective comparisons.

To this end the keras backend was imported to calculate the number of parameters in the model in _results_summary. This info was saved and assigned to the variable weight, which was then written to a csv in the results folder with _results_write, like all other information retrieved and returned by the _results_summary function. The string 'weights' was added to the 'metrics' variable in the 'results_comparison' function and a new function '_write_weight_table' was written corresponding to the _write_timetable' function, differing only in the addition of a manual offset...these functions can be combined in the future, just as it was done with the _write_MSE_MAE function. _write_weight_table was called at the appropriate position