Gen_trafos was made into a nested function of rescale and an extra trafo was added depending on the string value of the 'trafo' key in the features.cfg.
The gen_trafo function is only called if rescale is called, it makes no sense to have the rescale and gen_trafo functions to be completely independent of one another. For this reason gen_trafos is made into a nested function in rescale. Some further changes were made, so that this structural change in the code, caused no problems with scopes and definitons, these however were simple, are clear enough upon examining this commit. The most important change in this commit is perhaps the conditional choice of how the trafos are defined in gen_trafos depending on the value of the 'trafo' key, added in forecast.cfg. So far there is now a normalized scaling, which occurs if the trafo key in the configs is set to the string "norm" and a standardization scaling, which occurs if the aforementioned trafo key is set to the string "standard". It is important to note that the value inserted in the configs has to be a double quoted string. This is due to the fact that the configs parsed into the variable self.features passed into the rescale function are parsed with json.loads