Add apollo kpi and simplify nested functions in _result_comparsion.
The apollo kpi was added to result summary and the _result_comparison function was edited in order output this kpi to the xl document. This kpi measures the average distance of the hourly median with respect to the hourly maximum values (e.g. med_h1/pv_power_h1_max). This should ultimately be a measure of the ability of the network to learn the geographical dependencies of the solar modules output.

The various functions _wright_weight_table and _write_timetable were reduced to one function that fulfills both tasks.