resample features in parse features, remove half implemented code for multiple resolutions, calculate faith value.
There are two main changes to the code to discuss. First of all, multiple changes were made to the code so that features would be resampled only once in the parse features functions. All other calls of the resample method were removed from extract_inputs and extract_target methods defined in neural network and the respective network model classes. Furthermore, the cov method was defined, this method adds a new column labeled 'faith' this column contains the distance of the covariance between two features over the last 24 hrs from a population estimate of the covariance of these features, also calculated in this method. This difference is measured in units of the standard deviation of the 24hr covariance from the population covariance estimate. This method is called in _parse_features with 'dni' and 'pv_power' as the two features. This choice was motivated by a look at the distribution of these two features over the year, they seem to mirror eachother suggesting a large correlation between the two values. The choice of the name 'faith' for this new feature is because it should give us an idea as to how well the measurements on our module agree with the forecast. One would expect that a poor forecast would result in a poor correlation, as would a problem in the measurement apparatus.

It is extremely important to note that this update unvieled a fairly major, though easy to fix bug in th_e_simulation; that being that the network result is appended back to the unsampled features. This ultimately results in the result being resampled with actual measurements belonging to the hour that the networks is trying to predict. This ultimately explains why no difference was seen in the event horizons up until this point. This must be addressed in the master branch.