add necessary code to write doubt related kpi's to xl doc.
To allow for the evaluation of predictions based on the doubt values, added in the last commit to features, the doubt value must be recalculated everytime a prediction step is made by utilizing the last prediction for the next. The doubt values of each prediction are retrieved and ultimately added to the results dataframe generated by the simulate function, to be utilized later by the _results_summary function. in the _results_summary function the doubt values of each prediction are then used to reduce the error kpi's for larger doubt values.

In neural_network only small changes were made:

1. the average hourly doubt values of each hour were retrieved with the addition of the hourly_doubt method
2. the calculation of the standard deviation of the sample covariances from the population estimate was changed by dividing by the sample_size -1 instead of sample_size.