Fix: Order prediction correct time points for subsequent prediction in step_features + add timestep to date
Predictions made were previously ordered to the 60 minutes prior to the following timepoint to be predicted in the next iteration of the while step in step_index loop. This is incorrect because this predicted value belongs to the 60 minutes prior to the time just predicted. As is the prediction made for the previous iteration is not parsed for the inputs of the next iteration/step.

The other quick fix involves adding a timestep of one hour to date. This ensures that there are 72 hours before the very first timepoint to be predicted as well as the 60 min of info preceding the first prediction hour, which ultimately characterize the hour following the rolling resampling performed on the data.