Modify the dilation rate of the CNN layers.
This changes the dilation rate of the three cnn layers from (1, 4, 8) to (1, 2, 4) as it was in my bachelor thesis. A dilation rate of (1, 4, 8) for the three CNN's in the ConvLSTM indicates that each cell of information, after our input has passed through the cnn layers, is tasked with meaningfully extracting features from input spread over 24hrs. Tests showed that this was not feasible as training was rarely succesful. In comparison, a (1, 2, 4) CNN network is tasked with extracting features from information spread over 8 hrs. Training is succesful in the vast majority of cases.