Manually replace pv power in input corresponding to target hour with the average of the last hour.
Before this fix the last minute of the previous hour was inserted in the 60min corresponding to the target hour using the interpolation method. This resulted in one value corresponding to only 1/7320 of the data making up 1/73 of our input. Simply moving the interpolation step to after resampling was performed was not an option, as this estimate was required for the update of the doubt value.