Add attribute to the Resolution class called 'resolution'.
This attribute is added to the Resolution class and as such must have a corresponding entry in the model.cfg. This attribute signifies the time resolution of the database datapoints in minutes. This attribute is added to provide the user to make index translations at the smallest available time scale. This is important when resampling: For example when resampling the power output of a solar cell over the last hour the relevant data is in the index range [t- 59min, t] where t is the hour for which we are trying to calculate the average hour output. t-59 can now be expressed as t - time_step + resolution instead of writing t - time_step + dt.timedelta(minutes=1).