Rewrite code to illuminate that one input generates a 24 hour forecast.
The generation of a 24 hour forecast from a singular input is moved to the \_predict method in neural\_network. Exclusively the generation of input from features, as well as the cataloging of network predictions and targets for this generated input is carried out now in the \_simulate method. This was necessary to highlight the fact that one input generates 24 hour output. This clear method of generating input, mapping it a 24 hour vector with the network, and recording network the results should ease the evaluation of these network results in the future. For the timebeing outputs of this method have been removed and its purpose has truly been reduced to simulating predictions.