Remove and replace old kpi's with new ones.
The old kpi's are replaced with new ones. As of now the new kpi's consist of total mae, mbe, rmse, nrmse as well as the same kpi's calculated for individual doubt regions in the region characterized by the first forecast hour and the sun being higher then 35 degrees over the horizon.