Repackage my_evaluation into structure_err and evaluate methods.
The _parse_regions method was moved into the structure_err method, while the evaluate_regions method was moved to the evaluate regions. This restructuring ensures that the methods regarding categorizing our predictions into the various regions are all in one place, while the methods regarding evaluationg this categorized data is all in one place.