Restructure evalution around mi_kpi dataframe.
The previous evaluation used the preexisting functions used for the global evaluation prior to the multiindexing of results for each individual region. The output of this evaluation was far too verbose too be of any practical use. A more sensible approach was taken by using the mi_results object in order to calculate the mae, mbe, rmse, and nrmse for each unique index present in the multiindex of mi_results; this information is summarized in the object mi_kpis. All further evaluations of the network from now on should solely utilize this object (or derivatives of ti) and system info to achieve their output. In this commit, for example, the shadow function is introduced, in order to measure the ability of the network to predict the suppressive effect of morning and evening shadows on the pv_power output. It performs this task by calculating the meanmbe in all regions with a solar_elevation < 15 degrees.