Count number of points in each bin/region.
In bin_results the arbitrary value of 10 is set for the minimum number of points that can be placed in a bin. If less are found in any bin as the function runs, an error is raised and the program stops. The user is then asked to change the resolution/bin_width along the various axis of the feature space chosen for sorting. Furthermore a function in evaluation called bin_count is created as a sub function of the evaluation function. This function counts the number of points in each bin/region and stores this information in a pandas.Series object. This function is called in mi_kpi, and its output is concatenated as column of the final mi_kpi object. This should provide the user with an idea how much statistical values such as the relative kpi values can be trusted.