Modify how bin_count related outputs are handled
The bin results method no longer raises an error when 0 predictions are found corresponding to an index from the MultiIndex. Furthermore, in mi_kpi the output of bin_count method is cast to a pd.DataFrame whose columns are a MultiIndex object. This is to avoid the data in bin_count showing up as na values in the final mi_kpi DataFrame.