Remove the bin_count method.
The bin_count method was removed and its functionality was manually written in the necessary place in mi_kpi. This was done since having the method necessitates each kpi, to be potentially implemented in later commits, calls the bin_count method. This perhaps constrains our ability to generalize the kpi's coded in the future.

As a final note, the count data is copied once for each target, this is so that it is fully compatible with the MultiIndex of the mi_kpi DataFrame with the levels 'kpi' and 'targets'.