Modify shadows kpi.
A few minor modifications were made to the shadows kpi. Firstly, it is now required that pv_power be one of the values predicted by the network, as the utility of this kpi is for such predictions. Secondly the counting of all the points present in the final bins output by the kpi was implemented. Finally, the range of solar elevations taken into account was narrowed; it was found occasionally predictionsmade in the afternoon were being considered. The arguments present add_evaluation were slightly modified to handle the richer output of the shadows kpi (The value of 6 present in the loc function here is an artifact resulting from the dailypredictions made during the debugging process).