Calculate metrics along discrete valued axes utilizing the 'raw' results and without utilizing the mi.
It was realized that it is unecessary to estimate the error function before metric calculation, so long as the axis only takes on discrete values (opposed to continuous values). This has two main benefits:

1. Not needing to estimate the error function means we no longer have to group the results with regards to the discrete axes along which we would like to make an evaluation apriori. This avoids a bit of annoyance needing to check thatthe evaluation_data has the appropriate levels when it is loaded and if not to regroup the data according to the levels needed for the evaluation.

2. The stochastics of predictions made along these axes can be more accurately analyzed as there is no longer the intermediate averaging step of points together to create a regional estimate of the error function in a given region; each original prediction point is still available for the analysis.

There are a couple finapoints of note. Since we no longer have to check if the correct levels are available for the evaluation the discrete metrics are able to generalize to single function, called discrete metrics, whose behavior is configured by the evaluation.cfg. I would not claim however that a similar generalization would be impossible with the previous framework for calculation though, as previously mentioned, it would necessarily include the grouping and potential regrouping of the data. Also, the _print_boxplot function had to be slightly modified to be compatible with the workflow of the discrete_metrics function. The old _print_boxplot function is still available as _print_boxplot_mi as it assumes that the data is multiindexed. The old functions for calculating the various kpis are still present but not called; they will potentially be used in the future once the number of data points makes the lost stochastic due to the averaging process utilized to estimate the error function irrelevant.