Delete methods that discrete_metrics can perform.
To note is that the filter_data method is temporarily deleted in this commit. This is because this method is only called when grouping along a continuous axis. Until now grouping along a continuous axis, while possible, has not yet been implemnted in the evaluation configs. This is because many more use cases can be imagined for metrics grouped along discrete axes. This is due to the fact that humans can only divert attention to a handful of metrics in order to perform decision making. It is only once the evaluation has developed to the point that a computer can perform network evaluation and decide how to change the network parameters that metrics grouped around continuous axes will become of any use.