Remove filter parameter from configs.
Filtering the data will now be performed simply by manually adding the corresponding condition in the config file. Previously this was done automatically if the parameter was set to true in the evaluation config file. This was removed because methods for filtering the predictions of the neural network on other non-solar power targets have not yet been developed; in effect meaning a loss of generality was achieved if the config file contains the 'filter' option.