Reduce mi_results to discretization of continuous axes.
Coming up with a general grid to span the feature space, ultimately performs the same function as making chosen, continuous, axes discrete and using the the pandas groupby method to calculate statistics with respect to these discretized values. Implementing this drastically reduces the code, with mi_results being reduced to _extract_labels and being moved to be a subfunction of the evaluate function.

Currently the implementated discretation of an axis is performed by checking under the config subsection 'Discretized' to find key, value pairs representing a continuous axis and the number of discrete values this axis should be broken down into. The discrete axis is retrieved by taking even steps from the minimum value of the chosen feature in the config to the maximum value such that the total number of points retrieved is as specified in the config.