Remove independent Discretize section of the config
All changes in the code correspond to changes in the config file; that being the removal of the discretization section and the additionof the group_bins parameters in the individual evaluation sections.

As of this commit, if one wishes to group a given statistic (i.e. mae, mse, etc.) along one or more continuous axes (ex: pv_power), one is expected to add the parameter 'group_bins' to the evaluation and assign it a list of values (corresponding to the number of bins each continous axis should be grouped into). The order in which these values appear should correspond to the order they appear in the parameter 'groups'. If one only wishes to group along discrete axis (i.e hours, horizon, weekday) then this parameter should not appear in the evaluations configs. If one wishes to group along discrete and continuous axes then one should proceed as follows:

1. List the groups in the parameter groups with the discrete axes appearing first.

2. Add the parameter group_bins and assign it a list of values, with one value for each continuous axes.